refactor youtube check functions and rename files
